Different Concept Implementations/requirements.txt
Flask
Flask-Cors
sentence-transformers
scikit-learn
pyspellchecker
pandas
numpy
google-generativeai
